Move factory to the right place
