+++ src/factory/serialPortFactory.ts
